+++ st2tests/integration/mistral/test_wiring.py
@@ -510,3 +510,91 @@ class WiringTest(base.TestWorkflowExecution):
         # Wait for completion.
         execution = self._wait_for_completion(execution)
         self._assert_success(execution, num_tasks=2)
+
+    def test_pause_resume_cascade_to_subchain(self):
+        # A temp file is created during test setup. Ensure the temp file exists.
+        path = self.temp_dir_path
+        self.assertTrue(os.path.exists(path))
+
+        # Launch the workflow. The workflow will wait for the temp file to be deleted.
+        params = {'tempfile': path, 'message': 'foobar'}
+        action_ref = 'examples.mistral-test-pause-resume-subworkflow-chain'
+        execution = self._execute_workflow(action_ref, params)
+        execution = self._wait_for_task(execution, 'task1', 'RUNNING')
+
+        # Pause the workflow before the temp file is created. The workflow will be paused
+        # but task1 will still be running to allow for graceful exit.
+        execution = self.st2client.liveactions.pause(execution.id)
+
+        # Expecting the execution to be pausing, waiting for task1 to be completed.
+        execution = self._wait_for_state(execution, ['pausing'])
+
+        # Get the subworkflow execution.
+        task_executions = [e for e in self.st2client.liveactions.get_all()
+                           if e.context.get('parent', {}).get('execution_id') == execution.id]
+
+        subworkflow_execution = self.st2client.liveactions.get_by_id(task_executions[0].id)
+        subworkflow_execution = self._wait_for_state(subworkflow_execution, ['pausing'])
+
+        # Delete the temporary file.
+        os.remove(path)
+        self.assertFalse(os.path.exists(path))
+
+        # Wait for the executions to be paused.
+        subworkflow_execution = self._wait_for_state(subworkflow_execution, ['paused'])
+        execution = self._wait_for_state(execution, ['paused'])
+
+        # Resume the parent execution.
+        execution = self.st2client.liveactions.resume(execution.id)
+
+        # Wait for completion.
+        subworkflow_execution = self._wait_for_completion(subworkflow_execution)
+        self._assert_success(subworkflow_execution, num_tasks=2)
+
+        execution = self._wait_for_completion(execution)
+        self._assert_success(execution, num_tasks=2)
+
+    def test_pause_resume_cascade_subworkflow_from_chain(self):
+        # A temp file is created during test setup. Ensure the temp file exists.
+        path = self.temp_dir_path
+        self.assertTrue(os.path.exists(path))
+
+        # Launch the workflow. The workflow will wait for the temp file to be deleted.
+        params = {'tempfile': path, 'message': 'foobar'}
+        action_ref = 'examples.chain-test-pause-resume-with-subworkflow'
+        execution = self._execute_workflow(action_ref, params)
+
+        # Expecting the execution to be running.
+        execution = self._wait_for_state(execution, ['running'])
+
+        # Pause the workflow before the temp file is created. The workflow will be paused
+        # but task1 will still be running to allow for graceful exit.
+        execution = self.st2client.liveactions.pause(execution.id)
+
+        # Expecting the execution to be pausing, waiting for task1 to be completed.
+        execution = self._wait_for_state(execution, ['pausing'])
+
+        # Get the subworkflow execution.
+        task_executions = [e for e in self.st2client.liveactions.get_all()
+                           if e.context.get('parent', {}).get('execution_id') == execution.id]
+
+        subworkflow_execution = self.st2client.liveactions.get_by_id(task_executions[0].id)
+        subworkflow_execution = self._wait_for_state(subworkflow_execution, ['pausing'])
+
+        # Delete the temporary file.
+        os.remove(path)
+        self.assertFalse(os.path.exists(path))
+
+        # Wait for the executions to be paused.
+        subworkflow_execution = self._wait_for_state(subworkflow_execution, ['paused'])
+        execution = self._wait_for_state(execution, ['paused'])
+
+        # Resume the parent execution.
+        execution = self.st2client.liveactions.resume(execution.id)
+
+        # Wait for completion.
+        subworkflow_execution = self._wait_for_completion(subworkflow_execution)
+        self._assert_success(subworkflow_execution, num_tasks=2)
+
+        execution = self._wait_for_completion(execution)
+        self._assert_success(execution, num_tasks=2)